+++ Sample/src/minimum/java/com.live2d.demo.minimum/LAppMinimumLive2DManager.java
@@ -48,7 +48,7 @@ public class LAppMinimumLive2DManager {
 
         // 必要があればここで乗算する
         if (viewMatrix != null) {
-            projection.multiplyByMatrix(viewMatrix);
+            viewMatrix.multiplyByMatrix(projection);
         }
 
         model.update();