+++ Sample/src/minimum/java/com.live2d.demo.minimum/LAppMinimumModel.java
@@ -224,9 +224,14 @@ public class LAppMinimumModel extends CubismUserModel {
             LAppMinimumDelegate.getInstance().getActivity().finish();
         }
 
-        matrix.multiplyByMatrix(modelMatrix);
-        this.<CubismRendererAndroid>getRenderer().setMvpMatrix(matrix);
+        // キャッシュ変数の定義を避けるために、multiplyByMatrix()ではなく、multiply()を使用する。
+        CubismMatrix44.multiply(
+            modelMatrix.getArray(),
+            matrix.getArray(),
+            matrix.getArray()
+        );
 
+        this.<CubismRendererAndroid>getRenderer().setMvpMatrix(matrix);
         this.<CubismRendererAndroid>getRenderer().drawModel();
     }